Started properties, fixed bug, 3 dbls = jail
I started a properties class, not sure if I'm going about it the right
way.

Fixed a bug causing players to move an extra space after passing GO.

If a player rolls 3 doubles in a row they are then sent to jail and the
next player goes.
**NOTE - They are sent to jail BUT I did not add the ability to have to
escape from jail. As soon as it's their turn again, whatever they roll
they are free to move.