All kinds of Jail corrections and token change
I fixed jail to where it asks before you roll if you want to pay to get
out. There was another bug with staying in jail 3 turns. That is also
fixed. If you roll doubles to get out of jail you are not allowed to go
again for rolling them in accordance to rules. Also the if you don't
roll doubles by the time 3 turns is up you have to pay 50 bucks is also
in effect. All tokens now have the white around them deleted.